+++ src/input.rs
@@ -15,13 +15,13 @@ use uni_app::AppEvent;
 pub trait InputApi {
     // keyboard
     /// return the current status of a key (true if pressed)
-    fn key(&self, key: &str) -> bool;
+    fn key(&self, key: KeyEvent) -> bool;
     /// return true if a key was pressed since last update.
-    fn key_pressed(&mut self, key: &str) -> bool;
+    fn key_pressed(&mut self, key: KeyEvent) -> bool;
     /// return an iterator over all the keys that were pressed since last update.
     fn keys_pressed(&self) -> Keys;
     /// return true if a key was released since last update.
-    fn key_released(&mut self, key: &str) -> bool;
+    fn key_released(&mut self, key: KeyEvent) -> bool;
     /// return an iterator over all the keys that were released since last update.
     fn keys_released(&self) -> Keys;
     /// characters typed since last update
@@ -40,9 +40,9 @@ pub trait InputApi {
 }
 
 pub struct DoryenInput {
-    kdown: HashMap<String, bool>,
-    kpressed: HashMap<String, bool>,
-    kreleased: HashMap<String, bool>,
+    kdown: HashMap<KeyEvent, bool>,
+    kpressed: HashMap<KeyEvent, bool>,
+    kreleased: HashMap<KeyEvent, bool>,
     mdown: HashMap<usize, bool>,
     mpressed: HashMap<usize, bool>,
     mreleased: HashMap<usize, bool>,
@@ -75,16 +75,16 @@ impl DoryenInput {
             mouse_offset: (x_offset as f32, y_offset as f32),
         }
     }
-    fn on_key_down(&mut self, scan_code: &str) {
-        if !self.key(scan_code) {
-            self.kpressed.insert(scan_code.to_owned(), true);
-            self.kdown.insert(scan_code.to_owned(), true);
+    fn on_key_down(&mut self, scan_code: KeyEvent) {
+        if !self.key(scan_code.into()) {
+            self.kpressed.insert(scan_code, true);
+            self.kdown.insert(scan_code, true);
         }
     }
-    fn on_key_up(&mut self, scan_code: &str) {
-        self.kpressed.insert(scan_code.to_owned(), false);
-        self.kdown.insert(scan_code.to_owned(), false);
-        self.kreleased.insert(scan_code.to_owned(), true);
+    fn on_key_up(&mut self, scan_code: KeyEvent) {
+        self.kpressed.insert(scan_code, false);
+        self.kdown.insert(scan_code, false);
+        self.kreleased.insert(scan_code, true);
     }
     fn on_mouse_down(&mut self, button: usize) {
         if !self.mouse_button(button) {
@@ -108,10 +108,10 @@ impl DoryenInput {
     pub fn on_event(&mut self, event: &AppEvent) {
         match event {
             AppEvent::KeyDown(ref key) => {
-                self.on_key_down(&key.code);
+                self.on_key_down(keycode_to_key_event(&key.code));
             }
             AppEvent::KeyUp(ref key) => {
-                self.on_key_up(&key.code);
+                self.on_key_up(keycode_to_key_event(&key.code));
             }
             AppEvent::CharEvent(ch) => {
                 if !ch.is_control() {
@@ -149,19 +149,19 @@ impl DoryenInput {
 }
 
 impl InputApi for DoryenInput {
-    fn key(&self, scan_code: &str) -> bool {
-        matches!(self.kdown.get(scan_code), Some(&true))
+    fn key(&self, scan_code: KeyEvent) -> bool {
+        matches!(self.kdown.get(&scan_code), Some(&true))
     }
-    fn key_pressed(&mut self, scan_code: &str) -> bool {
-        matches!(self.kpressed.get(scan_code), Some(&true))
+    fn key_pressed(&mut self, scan_code: KeyEvent) -> bool {
+        matches!(self.kpressed.get(&scan_code), Some(&true))
     }
     fn keys_pressed(&self) -> Keys {
         Keys {
             inner: self.kpressed.iter().filter(|&(_, &v)| v),
         }
     }
-    fn key_released(&mut self, scan_code: &str) -> bool {
-        matches!(self.kreleased.get(scan_code), Some(&true))
+    fn key_released(&mut self, scan_code: KeyEvent) -> bool {
+        matches!(self.kreleased.get(&scan_code), Some(&true))
     }
     fn keys_released(&self) -> Keys {
         Keys {
@@ -189,7 +189,7 @@ impl InputApi for DoryenInput {
 }
 
 type KeyMapFilter<'a> =
-    Filter<std::collections::hash_map::Iter<'a, String, bool>, fn(&(&'a String, &'a bool)) -> bool>;
+    Filter<std::collections::hash_map::Iter<'a, KeyEvent, bool>, fn(&(&'a KeyEvent, &'a bool)) -> bool>;
 
 /// An iterator visiting all keys in arbitrary order.
 pub struct Keys<'a> {
@@ -197,9 +197,168 @@ pub struct Keys<'a> {
 }
 
 impl<'a> Iterator for Keys<'a> {
-    type Item = &'a str;
+    type Item = &'a KeyEvent;
 
     fn next(&mut self) -> Option<Self::Item> {
-        self.inner.next().map(|(k, _)| k.as_ref())
+        self.inner.next().map(|(k, _)| k)
+    }
+}
+
+#[derive(Debug, Clone, Copy, Hash, PartialEq, Eq)]
+pub enum KeyEvent {
+    Digit(u32),
+    Key(char),
+    
+    Esc,
+
+    F1,
+    F2,
+    F3,
+    F4,
+    F5,
+    F6,
+    F7,
+    F8,
+    F9,
+    F10,
+    F11,
+    F12,
+
+    // PrintScreen for windows
+    Snapshot,
+    ScrollLock,
+
+    Insert,
+    Home,
+    Del,
+    End,
+
+    PageDown,
+    PageUp,
+
+    ArrowLeft,
+    ArrowUp,
+    ArrowRight,
+    ArrowDown,
+
+    Backspace,
+    Enter,
+    Space,
+
+    NumLock,
+
+    // Could be Numpad(u32) so Numpad(num) but all numpad numbers except 5 are the same as ... so it would be a bit pointless
+    // Numpad0: Same as Insert
+    // Numpad1: Same as End
+    // Numpad2: Same as ArrowDown
+    // Numpad3: Same as PageDown
+    // Numpad4: Same as ArrowLeft
+    Numpad5,
+    // Numpad6: Same as ArrowRight
+    // Numpad7: Same as Home
+    // Numpad8: Same as ArrowUp
+    // Numpad9: Same as PageUp
+    NumpadAdd,
+    NumpadSub,
+    NumpadDivide,
+    
+    Backslash,
+    Caps,
+    Colon,
+    Comma,
+
+    Equal,
+    Backquote,
+
+    LBracket,
+    RBracket,
+
+    // LeftControl and RightControl both give back ControlLeft as `keycode` so these are both Ctrl
+    Ctrl,
+    LAlt,
+
+    LShift,
+    RShift,
+
+    Win,
+    Minus,
+    Period,
+
+    Tab,
+
+    // For keycodes in that translate to ""
+    // https://github.com/unrust/uni-app/blob/41246b070567e3267f128fff41ededf708149d60/src/native_keycode.rs#L160
+    Unknown,
+}
+
+fn keycode_to_key_event(code: &str) -> KeyEvent {
+    println!("KeyCode: {}", code);
+
+    match code {
+        c if c.contains("Digit") => {
+            let c = char::from(c.as_bytes()[5] as char);
+            KeyEvent::Digit(c.to_digit(10).unwrap())
+        },
+        c if c.contains("Key") => {
+            let c = char::from(c.as_bytes()[3] as char);
+            KeyEvent::Key(c.to_ascii_lowercase())
+        },
+
+        "Escape" => KeyEvent::Esc,
+
+        "F1" => KeyEvent::F1,
+        "F2" => KeyEvent::F2,
+        "F3" => KeyEvent::F3,
+        "F4" => KeyEvent::F4,
+        "F5" => KeyEvent::F5,
+        "F6" => KeyEvent::F6,
+        "F7" => KeyEvent::F7,
+        "F8" => KeyEvent::F8,
+        "F9" => KeyEvent::F9,
+        "F10" => KeyEvent::F10,
+        "F11" => KeyEvent::F11,
+        "F12" => KeyEvent::F12,
+
+        "ArrowUp" => KeyEvent::ArrowUp,
+        "ArrowDown" => KeyEvent::ArrowDown,
+        "ArrowLeft" => KeyEvent::ArrowLeft,
+        "ArrowRight" => KeyEvent::ArrowRight,
+
+        "Backspace" => KeyEvent::Backspace,
+        "Enter" => KeyEvent::Enter,
+        "Space" => KeyEvent::Space,
+
+        "PageUp" => KeyEvent::PageUp,
+        "PageDown" => KeyEvent::PageDown,
+
+        "NumLock" => KeyEvent::NumLock,
+        "Numpad5" => KeyEvent::Numpad5,
+        "NumpadAdd" => KeyEvent::NumpadAdd,
+        "NumpadSub" => KeyEvent::NumpadSub,
+        "NumpadDivide" => KeyEvent::NumpadDivide,
+
+        "Backslash" => KeyEvent::Backslash,
+        "CapsLock" => KeyEvent::Caps,
+        "Colon" => KeyEvent::Colon,
+        "Comma" => KeyEvent::Comma,
+        
+        "Equal" => KeyEvent::Equal,
+        "Backquote" => KeyEvent::Backquote,
+
+        "BracketLeft" => KeyEvent::LBracket,
+        "BracketRight" => KeyEvent::RBracket,
+
+        "ControlLeft" => KeyEvent::Ctrl,
+        "AltLeft" => KeyEvent::LAlt,
+
+        "ShiftLeft" => KeyEvent::LShift,
+        "ShiftRight" => KeyEvent::RShift,
+
+        "Win" => KeyEvent::Win,
+        "Minus" => KeyEvent::Minus,
+        "Period" => KeyEvent::Period,
+
+        "Tab" => KeyEvent::Tab,
+        _ => KeyEvent::Unknown,
     }
 }